Add a CI check for code formatting (#151)
This PR adds a GitHub action (🎉) that checks for the latest refmt syntax.

This is done thanks to @thomsj which brought up the problems with formatting in the repo! 👏 #105 (comment)

The current implementation is not ideal: it installs esy and then installs OCaml and @esy-ocaml/reason. A better way to do it is to distribute refmt as a standalone binary from the latest Reason release. This will be fast to install and therefore fast to integrate with CI.

An approach to do so is to add an artifact to the CircleCI runs, and then use something like circleci-artifacts.now.sh to fetch the artifact, extract it and profit 💰

In the meantime, it takes <4m to run it, which is much faster than the Azure Linux build, not to mention the Windows one.