DirectoryPortal: move files instead of copying
The fix in #275 introduced a bug that made `npm` unusable: `dircpy`
isn't aware of symlinks, so when it copied the files recursively, it
copied the actual file instead of keeping it a symlink. This made
`bin/npm` the actual JS file instead of being a symlink to
`../lib/node_modules/npm/whatever`.

This fixes #283