feat: adds `--arch` option to `fnm install`
fnm defaults to downloading the Node architecture that matches the
current system, a good default.

This commits adds an option to override that default where emulating
other architectures is possible and desired, to wit: MacOS 11 is arm64
but is capable of running x86_64 binaries, and developers may desire
running older versions of Node that are unavailble in arm64 form.