Allow symlinking to the system version
this PR allows `fnm alias system any_alias` and specifically `fnm default system`, for those who are seeking for it.
it also reveals that we need to do some housekeeping and clean the code. it is very complicated and I believe I can make it nicer. but this is for the future.