Allow recursive version lookups
Adding the ability to provide `--version-file-strategy=recursive` on
both `fnm env` and on each command can allow us to support recursive
resolutions of the Node versions.

This makes the default Node version as easy as:

```bash
echo "16" > ~/.node-version
```

(assuming your code lives inside your home directory) and fixes #442, if
becomes the default.

---

This fixes #144 which is a very old issue :smiley:
https://github.com/Schniz/fnm/issues/144#issuecomment-770798955