Don't warn on `~/.fnm` yet
As mentioned in https://github.com/Schniz/fnm/issues/357#issuecomment-1002106398,
we still install into ~/.fnm in the installation scripts. So in order to
not show warning on new installations, we need to defer the warning
until we have changed the installation script.

This means that we will support the new path :)